AdminTool internals refactorings
This makes it slightly easier to test command providers, and also aligns the visibilities.
For instance, methods that returned `Result` objects were publicly visible, while the `Result` interface was private.
