+++ community/command-line/src/main/java/org/neo4j/commandline/admin/AdminTool.java
@@ -23,12 +23,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.NoSuchElementException;
-import java.util.Objects;
-import java.util.function.Function;
 import java.util.function.Supplier;
-
-import org.neo4j.helpers.Service;
-import org.neo4j.helpers.collection.Iterables;
 
 import static java.lang.String.format;
 
@@ -41,9 +36,9 @@ public class AdminTool
         String extraHelp = System.getenv( "NEO4J_EXTRA_HELP" );
         boolean debug = System.getenv( "NEO4J_DEBUG" ) != null;
 
-        new AdminTool( new ServiceCommandLocator(), System.out::println, extraHelp, debug )
-                .execute( homeDir, configDir, args )
-                .exit();
+        AdminTool tool = new AdminTool( CommandLocator.fromServiceLocator(), System.out::println, extraHelp, debug );
+        Result result = tool.execute( homeDir, configDir, args );
+        result.exit();
     }
 
     private final String scriptName = "neo4j-admin";
@@ -54,7 +49,7 @@ public class AdminTool
 
     public AdminTool( CommandLocator locator, Output out, String extraHelp, boolean debug )
     {
-        this.locator = new AppendingLocator( help(), locator );
+        this.locator = CommandLocator.withAdditionalCommand( help(), locator );
         this.out = out;
         this.debug = debug;
         this.usage = new Usage( scriptName, out, this.locator, extraHelp );
@@ -152,55 +147,5 @@ public class AdminTool
     private static Result success()
     {
         return () -> System.exit( 0 );
-    }
-
-    private interface Result
-    {
-        void exit();
-    }
-
-    public interface CommandLocator
-            extends Function<String, AdminCommand.Provider>, Supplier<Iterable<AdminCommand.Provider>>
-    {
-    }
-
-    private static class ServiceCommandLocator implements CommandLocator
-    {
-        @Override
-        public AdminCommand.Provider apply( String name )
-        {
-            return Service.load( AdminCommand.Provider.class, name );
-        }
-
-        @Override
-        public Iterable<AdminCommand.Provider> get()
-        {
-            return Service.load( AdminCommand.Provider.class );
-        }
-    }
-
-    private class AppendingLocator implements CommandLocator
-    {
-        private final Supplier<AdminCommand.Provider> command;
-        private final CommandLocator commands;
-
-        public AppendingLocator( Supplier<AdminCommand.Provider> command, CommandLocator commands )
-        {
-            this.command = command;
-            this.commands = commands;
-        }
-
-        @Override
-        public AdminCommand.Provider apply( String name )
-        {
-            AdminCommand.Provider provider = command.get();
-            return Objects.equals( name, provider.name() ) ? provider : commands.apply( name );
-        }
-
-        @Override
-        public Iterable<AdminCommand.Provider> get()
-        {
-            return Iterables.append( command.get(), commands.get() );
-        }
     }
 }
